+++ SIM7000G/sim7000g.c
@@ -46,7 +46,7 @@ void connectToLTE(void) {
     vTaskDelay(pdMS_TO_TICKS(1000));
 }
 
-void sendDisplayData(float speed, int powerTotal, int powerIn, int powerOut, int battery, int motortemp, int mcutemp, int elevatorangle, double lattitude, double longitude,float rpm, float lowCelVoltage) {
+void sendDisplayData(float speed, int powerTotal, int powerIn, int powerOut, int battery, int motortemp, int mcutemp, int elevatorangle, double lattitude, double longitude,int rpm, float lowCelVoltage) {
 
     //sendATCommand("AT+SHCONF=\"URL\",\"http://server.domain.org\"",buffer);
     sendATCommand("AT+SHCONF=\"BODYLEN\",1024",buffer);
@@ -60,7 +60,7 @@ void sendDisplayData(float speed, int powerTotal, int powerIn, int powerOut, int
     sendATCommand("AT+SHAHEAD=\"Accept\",\"*/*\"",buffer);
 
     char data[150] = {0};
-    sprintf(data,"AT+SHREQ=\"/php/insert_data.php?s=%0.1f&p=%d&pi=%d&po=%d&b=%d&mt=%d&mct=%d&ea=%d&lat=%lf&lon=%lf&rpm=%d&lcv=%f\",3",speed,powerTotal,powerIn,powerOut,battery,motortemp,mcutemp, elevatorangle,lattitude,longitude,(int)rpm,lowCelVoltage);
+    sprintf(data,"AT+SHREQ=\"/php/insert_data.php?s=%0.1f&p=%d&pi=%d&po=%d&b=%d&mt=%d&mct=%d&ea=%d&lat=%lf&lon=%lf&rpm=%d&lcv=%f\",3",speed,powerTotal,powerIn,powerOut,battery,motortemp,mcutemp, elevatorangle,lattitude,longitude,rpm,lowCelVoltage);
     sendATCommand(data,buffer);
     vTaskDelay(pdMS_TO_TICKS(100));
     sendATCommand("AT+SHDISC",buffer);
